requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.1
packaging==23.0
python-docx==0.8.11
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
